Add a simplistic iCalendar file
This creates a highly minimal ics file which users can subscribe to
for a listing of our events.

An extension for the future would be to add the content of the event
page as the "description" of the event in question.

Things we could add to this, if we had the structured data:
 - Organiser
 - End time

Change-Id: I7ef61ae95710436ec6006ab60e686dcdc5ac2601