Add timezones to all our events
Jekyll's handling of dates seems to be a bit odd:
- if you don't specify a timezone, then it treats the date as
  having been specified in UTC. This is why the previous config
  (where the site timezone was UTC) was working fine.
- if you specifying a timezone it doesn't recognise (for example
  'Europe/London' or 'BST') then it treats the value as being in
  the site timezone
- if you specify a fixed numeric timezone (such as '+1') then it
  uses that timezone and converts to the site timezone for display

The second of these feels like it might actually be a bug, however
it is somewhat useful as we can just drop in Europe/London and be
done with it.

At the same time we also configure the site to render in our main
timezone. The net result is that the displayed times don't change
(I've checked this via `diff`), though we do now get the correct
timezones if we want them.