Rename sponsor-logo class to avoid adblock
It feels reasonable that we do this as these aren't advertisers --
they are genuinely sponsors of us as a charity, so aren't the kind
of thing that adblock extensions would normally aim to hide.
Additionally these content are served entirely from our own hosting,
so there's no user-tracking implications from these being shown.