Add subresource integrity checking to our user-tracking script
If we're going to track users we should at least ensure that the
script which does so is what we're expecting. The value here is
derived from a manual check of the current version of the file.

I couldn't find anything on Plausible's website about this, so
I'm just guessing that this is ok to do. Ideally it feels like
we should pin the version of the library we're pulling in, however
I couldn't find anything on how to do that either.