Remove competition pages link from the homepage
While we could make this configurable, we don't know what we'll
want this to say next year given that it was a new thing this year.
Removal is therefore the simplest way to decouple the comp pages
(thus enabling the shutdown of the srcomp machine) and it's still
easy to get this back from the history if we do want it.