Update HTML-proofer and pin the current major version
Version 4 introduces some breakage I don't have time to fix right
now and version 5 depends on Ruby 3. We should get to that eventually
however for now this is just unblocking other updates.

The main change here is that we're dropping the depedency on
`nokogumbo` which has been merged into `nokogiri` and isn't supported
on Ruby 3.