Update and pin Jekyll
Now that we're not building with GitHub Pages directly, we have more
freedom to update. That also means we need to be more explicit about
the versions we support. Jekyll version 4 is out, however seems to
have a rather different opinion of what our site looks like. Since
version 3 seems to still be somewhat supported we can keep using that
for now while we make other updates.